% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_taxa.R
\docType{data}
\name{all_taxa}
\alias{all_taxa}
\title{A table of all the the species found in FishBase, including taxonomic
classification and the Species Code (SpecCode) by which the species is
identified in FishBase.}
\description{
A table of all the the species found in FishBase, including taxonomic
classification and the Species Code (SpecCode) by which the species is
identified in FishBase.
}
\author{
Carl Boettiger \email{carl@ropensci.org}
}
\references{
\url{FishBase.org}
}
\keyword{data}

