% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{c_code}
\alias{c_code}
\title{c_code}
\usage{
c_code(
  c_code = NULL,
  server = getOption("FISHBASE_API", "fishbase"),
  version = get_latest_release(),
  db = default_db(),
  ...
)
}
\arguments{
\item{c_code}{a C_Code or list of C_Codes (FishBase country code)}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See \code{\link[=available_releases]{available_releases()}} for details.}

\item{db}{database connection, now deprecated.}

\item{...}{additional arguments, currently ignored}
}
\description{
return a table of country information for the requested c_code, as reported in FishBASE.org
}
\examples{
\dontshow{if (interactive() ) withAutoprint(\{ # examplesIf}
c_code(440)
\dontshow{\}) # examplesIf}
}
