% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_tbl.R
\name{clear_cache}
\alias{clear_cache}
\title{Clear rfishbase cache}
\usage{
clear_cache()
}
\description{
rfishbase caches data downloads for faster access.  Use this to reset 
the cache when changing versions of rfishbase.
}
