% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_names.R
\name{common_names}
\alias{common_names}
\alias{sci_to_common}
\title{common names}
\usage{
common_names(
  species_list = NULL,
  server = c("fishbase", "sealifebase"),
  version = "latest",
  db = NULL,
  Language = "English",
  fields = NULL
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See \code{\link[=available_releases]{available_releases()}} for details.}

\item{db}{database connection, now deprecated.}

\item{Language}{a string specifying the language for the common name, e.g. "English"}

\item{fields}{subset to these columns.  (recommend to omit this and handle manually)}
}
\value{
a data.frame of common names by species queried. If multiple species are queried,
The resulting data.frames are concatenated.
}
\description{
Return a table of common names
}
\details{
Note that there are many common names for a given sci name
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
common_names("Bolbometopon muricatum")
\dontshow{\}) # examplesIf}
}
