% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/commonnames.R
\name{commonnames}
\alias{commonnames}
\title{commonnames}
\usage{
commonnames(species_list, limit = 1000, server = SERVER, Language = NULL,
  fields = c("ComName", "Language", "C_Code", "SpecCode"))
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species").}

\item{limit}{The maximum number of matches from a single API call (e.g. per species). Function
will warn if this needs to be increased, otherwise can be left as is.}

\item{server}{base URL to the FishBase API, should leave as default.}

\item{fields}{a character vector specifying which fields (columns) should be returned. By default,
all available columns recognized by the parser are returned. This option can be used to the amount
of data transfered over the network if only certain columns are needed.}
}
\value{
a data.frame of common names by species queried. If multiple species are queried,
The resulting data.frames are concatenated.
}
\description{
Return a table of common names
}
\details{
Note that there are many common names for a given sci name, so sci_to_common doesn't make sense
}
\examples{
\donttest{
commonnames(c("Labroides bicolor",  "Bolbometopon muricatum"))

# subset by English language names
fish <- commonnames("Bolbometopon muricatum")
library(dplyr)
fish \%>\% filter(Language=="English")
}
}

