% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{countrysubref}
\alias{countrysubref}
\title{countrysubref}
\usage{
countrysubref(
  server = getOption("FISHBASE_API", "fishbase"),
  version = get_latest_release(),
  db = default_db(),
  ...
)
}
\arguments{
\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See \code{\link[=available_releases]{available_releases()}} for details.}

\item{db}{database connection, now deprecated.}

\item{...}{additional arguments, currently ignored}
}
\description{
return a table of countrysubref
}
\examples{
\dontshow{if (interactive() ) withAutoprint(\{ # examplesIf}
\dontrun{
countrysubref()
}
\dontshow{\}) # examplesIf}
}
