% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{db_delete}
\alias{db_delete}
\title{Delete the existing local fishbase database}
\usage{
db_delete(dbdir = db_dir())
}
\arguments{
\item{dbdir}{Path to the database.}
}
\description{
Delete the existing local fishbase database
}
