% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{db_disconnect}
\alias{db_disconnect}
\title{disconnect the database}
\usage{
db_disconnect(db = NULL)
}
\arguments{
\item{db}{optional, an existing pointer to the db, e.g. from [fb_conn()]
or [fb_import()].}
}
\description{
disconnect the database
}
