% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{faoareas}
\alias{faoareas}
\title{faoareas}
\usage{
faoareas(
  species_list = NULL,
  fields = NULL,
  server = getOption("FISHBASE_API", "fishbase"),
  version = "latest",
  db = NULL,
  ...
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{fields}{subset to these columns.  (recommend to omit this and handle manually)}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See \code{\link[=available_releases]{available_releases()}} for details.}

\item{db}{database connection, now deprecated.}

\item{...}{additional arguments, currently ignored}
}
\value{
a tibble, empty tibble if no results found
}
\description{
return a table of species locations as reported in FishBASE.org FAO location data
}
\examples{
\dontshow{if (interactive() ) withAutoprint(\{ # examplesIf}
  faoareas()
\dontshow{\}) # examplesIf}
}
