% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{fb_conn}
\alias{fb_conn}
\title{Cacheable database connection}
\usage{
fb_conn(
  server = c("fishbase", "sealifebase"),
  version = "latest",
  dir = db_dir()
)
}
\arguments{
\item{server}{fishbase or sealifebase}

\item{version}{release version}
}
\description{
Cacheable database connection
}
