% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_tbl.R
\name{fb_tables}
\alias{fb_tables}
\title{List the tables available on fishbase/sealifebase}
\usage{
fb_tables(server = c("fishbase", "sealifebase"), version = "latest")
}
\arguments{
\item{server}{Access data from fishbase or sealifebase?}

\item{version}{Version, see \code{\link[=available_releases]{available_releases()}}}
}
\description{
These table names can be used to access each of the corresponding tables
using \verb{[fb_tbl()]}.  Please note that following RDB design, it is often
necessary to join multiple tables.  Other data cleaning steps are sometimes
necessary as well.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
fb_tables()
\dontshow{\}) # examplesIf}
}
