% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio_pane.R
\name{fishbase_pane}
\alias{fishbase_pane}
\title{Open database connection pane in RStudio}
\usage{
fishbase_pane()
}
\description{
This function launches the RStudio "Connection" pane to interactively
explore the database.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

if (!is.null(getOption("connectionObserver"))) fishbase_pane()
\dontshow{\}) # examplesIf}
}
