% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getFaoArea}
\alias{getFaoArea}
\title{get FAO area table}
\usage{
getFaoArea(fish.data = NULL, path = NULL)
}
\arguments{
\item{fish.data}{the fishbase database fish.data or a subset,}

\item{path}{to cached copy of fishbase (optional, defaults to copy in package).}
}
\value{
A list of data frames, one per species.
}
\description{
return the FAO data for the required species
}
\examples{
data(fishbase)
out <- getFaoArea(fish.data[1:3])

# or using species names:
ids <- findSpecies(c("Coris_pictoides", "Labropsis_australis"))
out <- getFaoArea(fish.data[ids])
}

