% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dynamics.R
\name{length_length}
\alias{length_length}
\alias{popll}
\title{length_length}
\usage{
length_length(
  species_list = NULL,
  fields = NULL,
  server = c("fishbase", "sealifebase"),
  version = "latest",
  db = NULL,
  ...
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{fields}{subset to these columns.  (recommend to omit this and handle manually)}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See \code{\link[=available_releases]{available_releases()}} for details.}

\item{db}{database connection, now deprecated.}

\item{...}{additional arguments, currently ignored}
}
\value{
a table of lengths
}
\description{
return a table of lengths
}
\details{
This table contains relationships for the conversion of one length type to another for over 8,000
species of fish, derived from different publications, e.g. Moutopoulos and Stergiou (2002) and
Gaygusuz et al (2006), or from fish pictures, e.g. Collette and Nauen (1983), Compagno (1984)
and Randall (1997). The relationships, which always refer to centimeters, may consist either of a
regression linking two length types, of the form:
Length type (2) = a + b x Length type (1)
Length type (2) = b' x Length type (1)
The available length types are, as elsewhere in FishBase,
TL = total length;
FL = fork length;
SL = standard length;
WD = width (in rays);
OT = other type (to be specified in the Comment field).
When a version of equation (1) is presented, the length range, the number of fish used in the regression,
the sex and the correlation coefficient are presented, if available.
When a version of equation (2) is presented, the range and the correlation coefficient are omitted,
as the ratio in (2) will usually be estimated from a single specimen, or a few fish covering a narrow
range of lengths.

\strong{Note on column naming:} The columns \code{Length1} and \code{Length2} follow FishBase's internal
database convention, which may appear counterintuitive. \code{Length1} is the \emph{unknown} length
being predicted, and \code{Length2} is the \emph{known} length used as the predictor. The formula
is therefore: \code{Length1 = a + b * Length2}. This matches the FishBase website, which labels
the columns "Unknown length" and "Known length" respectively.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
\dontrun{
length_length("Oreochromis niloticus")
}
\dontshow{\}) # examplesIf}
}
\references{
http://www.fishbase.org/manual/english/PDF/FB_Book_CBinohlan_Length-Length_RF_JG.pdf
}
