% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxa.R
\name{load_taxa}
\alias{load_taxa}
\title{load_taxa}
\usage{
load_taxa(update = FALSE, cache = TRUE, server = getOption("FISHBASE_API",
  FISHBASE_API), version = NULL, limit = 5000L, ...)
}
\arguments{
\item{update}{logical, should we query the API to update the available list?}

\item{cache}{should we cache the updated version throughout this session? 
(default TRUE, leave as is)}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase}

\item{version}{API version, only applies to Fishbase right now. See 
\code{\link{versions}} to get versions available.}

\item{limit}{The maximum number of matches from a single API call (e.g. per species). Function
will warn if this needs to be increased, otherwise can be left as is.}

\item{...}{additional arguments to \code{\link[httr]{GET}}}
}
\value{
the taxa list
}
\description{
Load or update the taxa list
}
