% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{locations}
\usage{
locations(species_list, server = SERVER, verbose = TRUE, limit = 100)
}
\arguments{
\item{species_list}{}
}
\description{
return a table of species locations as reported in FishBASE.org FAO location data
}

