% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{occurrence}
\alias{occurrence}
\title{occurrence}
\usage{
occurrence()
}
\description{
occurrence
}
\details{
THE OCCURRENCE TABLE HAS BEEN DROPPED BY FISHBASE - THIS
FUNCTION NOW RETURNS A STOP MESSAGE.
}

