% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase.R
\docType{package}
\name{rfishbase-package}
\alias{rfishbase-package}
\alias{rfishbase}
\title{The new R interface to Fishbase, v5.0}
\description{
A programmatic interface to FishBase, re-written
based on an accompanying 'RESTful' API. Access tables describing over 30,000
species of fish, their biology, ecology, morphology, and more. This package also
supports experimental access to SeaLifeBase data, which contains
nearly 200,000 species records for all types of aquatic life not covered by
FishBase.'
}
\section{A message from the FishBase team}{

Thank you for using FishBase and SeaLifeBase through the rfishbase package,
representing more than 35 and 20 years of efforts, respectively. With
databases of such a large scope, both in terms of species and life traits
covered, one cannot expect that it is complete and free of errors. We would
be grateful if you could spare some of your time, to directly report to
FishBase and SeaLifeBase team:

\itemize{
\item mistakes, typos, taxonomic issues, etc.;
\item if you have completed the dataset on your own, by sending the pdfs
or reference citations that would allow us to address some missing data;
\item by sending articles that you have published using the data, pdfs or
reference citations.
}

Please send your comments and corrections to \email{fishbase@q-quatics.org}.
You may send PDF files directly to the same address.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/rfishbase/}
  \item \url{https://github.com/ropensci/rfishbase}
  \item Report bugs at \url{https://github.com/ropensci/rfishbase/issues}
}

}
\author{
Carl Boettiger \email{carl@ropensci.org}

Scott Chamberlain \email{scott@ropensci.org}
}
\keyword{package}
