% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/taxa.R
\name{sci_names}
\alias{sci_names}
\title{Get scientific names for the queried group}
\usage{
sci_names(query, all_taxa = load_taxa())
}
\description{
Get scientific names for the queried group
}
\examples{
sci_names(list(Family = 'Scaridae'))
sci_names(list(Genus = 'Labroides'))
}

