% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/load_taxa.R
\docType{data}
\name{sealifebase}
\alias{sealifebase}
\title{A table of all the the species found in SeaLifeBase, including taxonomic
classification and the Species Code (SpecCode) by which the species is
identified in SeaLifeBase}
\description{
A table of all the the species found in SeaLifeBase, including taxonomic
classification and the Species Code (SpecCode) by which the species is
identified in SeaLifeBase
}
\author{
Carl Boettiger \email{carl@ropensci.org}
}
\references{
\url{www.sealifebase.org}
}
\keyword{data}

