% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species.R
\docType{data}
\name{species_fields}
\alias{species_fields}
\title{A list of the species_fields available}
\description{
A list of the species_fields available
}
\references{
\url{http://fishbase.org}
}
\author{
Carl Boettiger \email{carl@ropensci.org}
}
\keyword{data}
