% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/species_info.R
\name{species_info}
\alias{species_info}
\title{species_info}
\usage{
species_info(species_list, fields = NULL, limit = 200, server = SERVER)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species").}

\item{fields}{a character vector specifying which fields (columns) should be returned. By default,
all available columns recognized by the parser are returned. This option can be used to the amount
of data transfered over the network if only certain columns are needed.}

\item{limit}{The maximum number of matches from a single API call (e.g. per species). Function
will warn if this needs to be increased, otherwise can be left as is.}

\item{server}{base URL to the FishBase API, should leave as default.}
}
\value{
a data.frame with rows for species and columns for the fields returned by the query (FishBase 'species' table)
}
\description{
Provide wrapper to work with species lists.
}
\details{
The Species table is the heart of FishBase. This function provides a convenient way to
query, tidy, and assemble data from that table given an entire list of species.
For details, see: http://www.fishbase.org/manual/english/fishbasethe_species_table.htm

Species scientific names are defined according to fishbase taxonomy and nomenclature.
}
\examples{
\donttest{

species_info(c("Labroides bicolor", "Bolbometopon muricatum"))
species_info(c("Labroides bicolor", "Bolbometopon muricatum"), fields = species_fields$habitat)

}
}

