% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/species_list.R
\name{species_list}
\alias{species_list}
\title{species_list}
\usage{
species_list(Class = NULL, Order = NULL, Family = NULL,
  SubFamily = NULL, Genus = NULL, Species = NULL, SpecCode = NULL,
  SpeciesRefNo = NULL, all_taxa = load_taxa())
}
\description{
Get a list of scientific names for all species in the given taxonomic group
}
\details{
The first time the function is called it will download and cache the complete
}
\examples{
\donttest{
## All species in the Family Scaridae
  species_list(Family = 'Scaridae')
## All species in the Genus
  species_list(Genus = 'Labroides')
}
}

