% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_names.R
\name{species_names}
\alias{species_names}
\title{species names}
\usage{
species_names(codes, server = getOption("FISHBASE_API", FISHBASE_API))
}
\arguments{
\item{codes}{a vector of speccodes (e.g. column from a table)}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase}
}
\value{
A character vector of species names for the SpecCodes
}
\description{
returns species names given FishBase's SpecCodes
}
