% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/species_table.R
\name{species_table}
\alias{species_table}
\title{species_table}
\usage{
species_table(species_list, verbose = TRUE, limit = 50, server = SERVER,
  fields = NULL)
}
\arguments{
\item{species_list}{Takes a vector of scientific names (each element as "genus species").
if only one name is given in an element (no space), assumes it is the genus and returns
all species matching that genus.}

\item{verbose}{should the function give warnings?}

\item{limit}{The maximum number of matches from a single API call.  (Applicable only when
querying at genus level, typically only one #' record will match the scientific name and
will try all names given.)  Thus the default is almost always fine.}

\item{server}{URL of the FishBase API server.}
}
\value{
a data.frame with rows for species and columns for the fields returned by the query (FishBase 'species' table)
}
\description{
Provide wrapper to work with species lists.
}
\details{
The Species table is the heart of FishBase. This function provides a convenient way to
query, tidy, and assemble data from that table given an entire list of species.
For details, see: http://www.fishbase.org/manual/english/fishbasethe_species_table.htm

Species scientific names are defined according to fishbase taxonomy and nomenclature.
}
\examples{
\donttest{

species_table(c("Oreochromis niloticus", "Bolbometopon muricatum"))
# There are 5 species in this genus, so returns 5 rows:
species_table("Labroides")
}
}

