% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/speciesnames.R
\name{speciesnames}
\alias{speciesnames}
\title{speciesnames}
\usage{
speciesnames(codes, all_taxa = load_taxa())
}
\arguments{
\item{codes}{a vector of speccodes (e.g. column from a table)}

\item{all_taxa}{the taxa table, usually loaded by default}
}
\value{
A character vector of species names for the SpecCodes
}
\description{
returns species names given FishBase's SpecCodes
}

