\docType{data}
\name{tree}
\alias{tree}
\title{An example phylogeny of some labrids.}
\description{
  An example phylogeny of some labrids.
}
\references{
  Price, S. a, Wainwright, P. C., Bellwood, D. R.,
  Kazancioglu, E., Collar, D. C., & Near, T. J. (2010).
  Functional innovations and morphological diversification
  in parrotfishes. Evolution; international journal of
  organic evolution, 3057-3068.
  doi:10.1111/j.1558-5646.2010.01036.x
}
\keyword{data}

