% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versions.R
\name{versions}
\alias{versions}
\title{Fishbase API versions}
\usage{
versions(server = getOption("FISHBASE_API", FISHBASE_API), ...)
}
\arguments{
\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase}

\item{...}{additional arguments to httr::GET}
}
\value{
a list
}
\description{
Fishbase API versions
}
\examples{
\dontrun{
versions()
}
}
