% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_tbl.R
\name{available_releases}
\alias{available_releases}
\title{List available releases}
\usage{
available_releases(server = c("fishbase", "sealifebase"))
}
\arguments{
\item{server}{fishbase or sealifebase}
}
\description{
List available releases
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
available_releases()
\dontshow{\}) # examplesIf}
}
