% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_tbl.R
\name{available_releases}
\alias{available_releases}
\title{List available releases}
\usage{
available_releases()
}
\description{
List available releases
}
\details{
Lists all available releases (year.month format).  
To use a specific release, set the desired release using
`options(FISHBASE_VERSION=)`, as shown in the examples. 
Otherwise, rfishbase will use the latest available version if this
option is unset.  NOTE: it will be necessary 
to clear the cache with `clear_cache()` or by restarting the R session
with a fresh environment.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
available_releases()
options(FISHBASE_VERSION="19.04")
## unset
options(FISHBASE_VERSION=NULL)
\dontshow{\}) # examplesIf}
}
