% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{c_code}
\alias{c_code}
\title{c_code}
\usage{
c_code(c_code = NULL, server = getOption("FISHBASE_API", FISHBASE_API), ...)
}
\arguments{
\item{c_code}{a C_Code or list of C_Codes (FishBase country code)}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase}

\item{...}{unused; for backwards compatibility only}
}
\description{
return a table of country information for the requested c_code, as reported in FishBASE.org
}
\details{
e.g. http://www.fishbase.us/Country
}
\examples{
\dontrun{
c_code(440)
}
}
