% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/commonnames.R
\name{common_to_sci}
\alias{common_to_sci}
\title{}
\usage{
common_to_sci(x, Language = NULL, verbose = TRUE, limit = 1000,
  server = SERVER)
}
\description{

}
\examples{
\donttest{
common_to_sci(c("Bicolor cleaner wrasse", "humphead parrotfish"))
common_to_sci("trout")
common_to_sci(c("trout", "Coho Salmon"))
}
}
\seealso{
\code{\link{commonnames}}, \code{\link{species_list}}, \code{\link{synonyms}}
}

