% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{countrysub}
\alias{countrysub}
\title{countrysub}
\usage{
countrysub(
  species_list = NULL,
  fields = NULL,
  server = c("fishbase", "sealifebase"),
  version = "latest",
  ...
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See [available_releases()] for details.}
}
\description{
return a table of countrysub for the requested species
}
\examples{
\dontshow{if (interactive() ) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
countrysub(species_list(Genus='Labroides'))
}
\dontshow{\}) # examplesIf}
}
