% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{countrysubref}
\alias{countrysubref}
\title{countrysubref}
\usage{
countrysubref(server = NULL)
}
\arguments{
\item{server}{can be set to either "fishbase" or "sealifebase" to switch between databases. NOTE: it is usually
easier to leave this as NULL and set the source instead using the environmental variable `FISHBASE_API`, e.g.
`Sys.setenv(FISHBASE_API="sealifebase")`.}
}
\description{
return a table of countrysubref
}
\examples{
\dontrun{
countrysubref()
}
}
