% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_ecology.R
\name{diet_items}
\alias{diet_items}
\title{fooditems}
\usage{
diet_items(
  server = getOption("FISHBASE_API", "fishbase"),
  version = get_latest_release(),
  db = default_db(),
  ...
)
}
\arguments{
\item{server}{can be set to either "fishbase" or "sealifebase" to switch between databases. NOTE: it is usually
easier to leave this as NULL and set the source instead using the environmental variable `FISHBASE_API`, e.g.
`Sys.setenv(FISHBASE_API="sealifebase")`.}

\item{version}{a version string for the database, will default to the latest release. see [get_releases()] for details.}

\item{db}{the}

\item{...}{unused; for backwards compatibility only}
}
\value{
a table of species diet_items
}
\description{
fooditems
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 \dontrun{
diet_items("Oreochromis niloticus")
}
\dontshow{\}) # examplesIf}
}
