% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_ecology.R
\name{diet_items}
\alias{diet_items}
\title{fooditems}
\usage{
diet_items(server = NULL)
}
\arguments{
\item{server}{can be set to either "fishbase" or "sealifebase" to switch between databases. NOTE: it is usually
easier to leave this as NULL and set the source instead using the environmental variable `FISHBASE_API`, e.g.
`Sys.setenv(FISHBASE_API="sealifebase")`.}
}
\value{
a table of species diet_items
}
\description{
fooditems
}
\examples{
\dontrun{
diet_items("Oreochromis niloticus")
}
}
