% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_ecology.R
\name{diet_items}
\alias{diet_items}
\title{diet_items}
\usage{
diet_items(...)
}
\arguments{
\item{...}{additional arguments (not used)}
}
\value{
a table of diet_items
}
\description{
diet_items
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
diet_items()
}
\dontshow{\}) # examplesIf}
}
