% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trophic_ecology.R
\name{ecology}
\alias{ecology}
\title{ecology}
\usage{
ecology(species_list = NULL, fields = NULL,
  server = getOption("FISHBASE_API", FISHBASE_API))
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{fields}{a character vector specifying which fields (columns) should be returned. By default,
all available columns recognized by the parser are returned.  Mostly for backwards compatibility as users can subset by column later}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase}
}
\value{
a table of species ecology data
}
\description{
ecology
}
\details{
By default, will only return one entry (row) per species.  Increase limit to
get multiple returns for different stocks of the same species, though often data is either
identical to the first or simply missing in the additional stocks.
}
\examples{
\dontrun{
ecology("Oreochromis niloticus")

## trophic levels and standard errors for a list of species
ecology(c("Oreochromis niloticus", "Salmo trutta"),
        fields=c("SpecCode", "FoodTroph", "FoodSeTroph", "DietTroph", "DietSeTroph"))
}
}
\references{
http://www.fishbase.org/manual/english/fishbasethe_ecology_table.htm
}
