% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ecology.R
\name{ecology}
\alias{ecology}
\title{ecology}
\usage{
ecology(species_list, server = SERVER, limit = 100, fields = NULL)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species").
(FishBase SpecCodes can be given as numeric values in place of a scientific name.)}

\item{server}{base URL to the FishBase API, should leave as default.}

\item{limit}{The maximum number of matches from a single API call (e.g. per species). Function
will warn if this needs to be increased, otherwise can be left as is.}

\item{fields}{a character vector specifying which fields (columns) should be returned. By default,
all available columns recognized by the parser are returned. This option can be used to the amount
of data transfered over the network if only certain columns are needed.}
}
\description{
return a table of species ecology as reported in FishBASE.org FAO location data
}
\examples{
\dontrun{
#' ecology("Oreochromis niloticus")
}
}

