% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{ecosystem}
\alias{ecosystem}
\title{ecosystem}
\usage{
ecosystem(
  species_list = NULL,
  fields = NULL,
  server = getOption("FISHBASE_API", "fishbase"),
  version = get_latest_release(),
  db = default_db(),
  ...
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See [available_releases()] for details.}
}
\value{
a table of species ecosystems data
}
\description{
ecosystem
}
\examples{
\dontshow{if (interactive() ) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
ecosystem("Oreochromis niloticus")
}
\dontshow{\}) # examplesIf}
}
