% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{faoareas}
\alias{faoareas}
\title{faoareas}
\usage{
faoareas(
  species_list = NULL,
  fields = NULL,
  server = getOption("FISHBASE_API", "fishbase"),
  version = "latest",
  db = NULL,
  ...
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See [available_releases()] for details.}
}
\value{
a tibble, empty tibble if no results found
}
\description{
return a table of species locations as reported in FishBASE.org FAO location data
}
\examples{
\dontshow{if (interactive() ) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
  faoareas()
}
e.g. http://www.fishbase.us/Country/FaoAreaList.php?ID=5537
\dontshow{\}) # examplesIf}
}
