% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locations.R
\name{faoareas}
\alias{faoareas}
\title{faoareas}
\usage{
faoareas(species_list, server = SERVER, limit = 500)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species").}

\item{server}{base URL to the FishBase API, should leave as default.}

\item{limit}{The maximum number of matches from a single API call (e.g. per species). Function
will warn if this needs to be increased, otherwise can be left as is.}
}
\description{
return a table of species locations as reported in FishBASE.org FAO location data
}
\details{
currently this is ~ FAO areas table (minus "note" field)
e.g. http://www.fishbase.us/Country/FaoAreaList.php?ID=5537
}
\examples{
\donttest{
  faoareas(species_list(Genus='Labroides'))
}
}

