% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_db.R
\name{fb_conn}
\alias{fb_conn}
\title{Cacheable database connection}
\usage{
fb_conn(server = c("fishbase", "sealifebase"), version = "latest")
}
\arguments{
\item{server}{fishbase or sealifebase}

\item{version}{release version}

\item{dir}{directory where database cache will be stored}
}
\description{
Cacheable database connection
}
