% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet_db.R
\name{fb_tables}
\alias{fb_tables}
\title{fb_tables
list tables}
\usage{
fb_tables(server = c("fishbase", "sealifebase"), version = "latest")
}
\arguments{
\item{server}{fishbase or sealifebase}

\item{version}{release version}
}
\description{
fb_tables
list tables
}
