% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_tbl.R
\name{fb_tbl}
\alias{fb_tbl}
\title{Access a fishbase or sealifebase table}
\usage{
fb_tbl(
  tbl,
  server = c("fishbase", "sealifebase"),
  version = "latest",
  db = fb_conn(server, version),
  collect = TRUE
)
}
\arguments{
\item{tbl}{table name, as it appears in the database. See [fb_tables()]
for a list.}

\item{server}{fishbase or sealifebase}

\item{version}{release version}

\item{db}{A cachable duckdb database connection}

\item{collect}{should we return an in-memory table? Generally best to leave
as TRUE unless RAM is too limited.  A remote table can be used with most
dplyr functions (filter, select, joins, etc) to further refine.}
}
\description{
Access a fishbase or sealifebase table
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
fb_tbl("species")
\dontshow{\}) # examplesIf}
}
