% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{fish.data}
\alias{fish.data}
\title{A cached copy of extracted FishBase data, 14 Febuary 2013.}
\description{
A cached copy of extracted FishBase data, 14 Febuary 2013.
}
\references{
Froese, R. and D. Pauly. Editors. 2011.FishBase.
  World Wide Web electronic publication.
  \url{www.fishbase.org}, version (10/2011).
}
\keyword{data}

