\docType{data}
\name{fish.data}
\alias{fish.data}
\title{The fishbase data set, cached.}
\description{
  The fishbase data set, cached.
}
\references{
  Froese, R. and D. Pauly. Editors. 2011.FishBase.  World
  Wide Web electronic publication.  \url{www.fishbase.org},
  version (10/2011).
}
\keyword{data}
\keyword{datasets}

