% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxa.R
\docType{data}
\name{fishbase}
\alias{fishbase}
\title{A table of all the the species found in FishBase, including taxonomic
classification and the Species Code (SpecCode) by which the species is
identified in FishBase.}
\description{
A table of all the the species found in FishBase, including taxonomic
classification and the Species Code (SpecCode) by which the species is
identified in FishBase.
}
\references{
\url{http://fishbase.org}
}
\author{
Carl Boettiger \email{carl@ropensci.org}
}
\keyword{data}
