% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{heartbeat}
\alias{heartbeat}
\title{heartbeat}
\usage{
heartbeat(server = SERVER)
}
\arguments{
\item{server}{base URL to the FishBase API, should leave as default.}
}
\value{
An httr 'response' object.
}
\description{
Check that the FishBase API server is responding
}
\examples{
\dontrun{

## Show server response times
resp <- heartbeat()
resp$times

## Show API endpoints:
content(resp)

}
}

