% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{introductions}
\alias{introductions}
\title{introductions}
\usage{
introductions(species_list = NULL, server = getOption("FISHBASE_API",
  FISHBASE_API), ...)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, the table will be loaded up to the limit. This makes it possible to load an entire table without knowledge of the species in the table, thus avoiding uneccesary API calls.}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase}

\item{...}{additional arguments to httr::GET}
}
\value{
a table of species introductions data
}
\description{
introductions
}
\examples{
\dontrun{
introductions("Oreochromis niloticus")
}
}
