% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{introductions}
\alias{introductions}
\title{introductions}
\usage{
introductions(
  species_list = NULL,
  fields = NULL,
  server = c("fishbase", "sealifebase"),
  version = "latest",
  ...
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{fields}{subset to these columns.  (recommend to omit this and handle manually)}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See [available_releases()] for details.}
}
\value{
a table of species introductions data
}
\description{
introductions
}
\examples{
\dontshow{if (interactive() ) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
introductions("Oreochromis niloticus")
}
\dontshow{\}) # examplesIf}
}
