% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_dynamics.R
\name{length_freq}
\alias{length_freq}
\alias{poplf}
\title{length_freq}
\usage{
length_freq(species_list = NULL, fields = NULL,
  server = getOption("FISHBASE_API", FISHBASE_API), ...)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{fields}{a character vector specifying which fields (columns) should be returned. By default,
all available columns recognized by the parser are returned.  Mostly for backwards compatibility as users can subset by column later}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase}

\item{...}{unused; for backwards compatibility only}
}
\value{
a table of length_freq information by species; see details
}
\description{
return a table of species fooditems
}
\examples{
\dontrun{
length_freq("Oreochromis niloticus")
}
}
\references{
http://www.fishbase.org/manual/english/lengthfrequency.htm
}
