% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_fields.R
\name{list_fields}
\alias{list_fields}
\title{List all fields (column names) in a set of tables}
\usage{
list_fields(
  import = FALSE,
  server = c("fishbase", "sealifebase"),
  version = "latest",
  db = fb_conn(server, version),
  tables = NULL
)
}
\arguments{
\item{import}{logical, default FALSE. Should we import missing tables first?}

\item{server}{fishbase or sealifebase}

\item{version}{release version}

\item{db}{A cachable duckdb database connection}

\item{tables}{list of tables of interest.  Missing tables will be imported
first if import=TRUE. 
Leave as NULL to query (and thus import) all tables (possibly slow!)}
}
\value{
a named list of length(tables). Each element of the list is named
for the corresponding table and contains a character vector listing the
fields (column names).
}
\description{
List all fields (column names) in a set of tables
}
