% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_taxa.R
\name{load_taxa}
\alias{load_taxa}
\title{load_taxa}
\usage{
load_taxa(update = FALSE, cache = TRUE, server = SERVER)
}
\arguments{
\item{update}{logical, should we query the API to update the available list?}

\item{cache}{should we cache the updated version throughout this session?
(default TRUE, leave as is)}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use http://fishbase.ropensci.org/sealifebase}
}
\value{
the taxa list
}
\description{
Load or update the taxa list
}

