% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxa.R
\name{load_taxa}
\alias{load_taxa}
\title{load_taxa}
\usage{
load_taxa(server = c("fishbase", "sealifebase"), version = "latest", ...)
}
\arguments{
\item{server}{Either "fishbase" (the default) or "sealifebase"}

\item{version}{the version of the database you want. Will default to the
latest available; see [available_releases()].}

\item{...}{for compatibility with previous versions}
}
\value{
the taxa list
}
\description{
load_taxa
}
