% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxa.R
\name{load_taxa}
\alias{load_taxa}
\title{load_taxa}
\usage{
load_taxa(server = getOption("FISHBASE_API", "fishbase"),
  version = get_latest_release(), db = default_db(), ...)
}
\arguments{
\item{server}{API for Fishbase or Sealifebase?}

\item{...}{for compatibility with previous versions}
}
\value{
the taxa list
}
\description{
load_taxa
}
