\name{loadCache}
\alias{loadCache}
\title{Load an updated cache}
\usage{
loadCache(path = NULL)
}
\arguments{
  \item{path}{location where cache is located}
}
\value{
loads the object fish.data into the working space.
}
\description{
Load an updated cache
}
\examples{
\dontrun{
 updateCache()
 loadCache()
}
}
\seealso{
\code{\link{updateCache}}
}
\keyword{cache}

