% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{locations}
\usage{
locations(species_list, server = SERVER, verbose = TRUE, limit = 100)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species").
(FishBase SpecCodes can be given as numeric values in place of a scientific name.)}

\item{server}{base URL to the FishBase API, should leave as default.}

\item{verbose}{logical. Should the function give warnings? (default TRUE)}

\item{limit}{The maximum number of matches from a single API call (e.g. per species).  When verbose=TRUE, function
will warn if this needs to be increased, otherwise can be left as is.}
}
\description{
return a table of species locations as reported in FishBASE.org FAO location data
}

