% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproduction.R
\name{maturity}
\alias{maturity}
\title{maturity}
\usage{
maturity(
  species_list = NULL,
  fields = NULL,
  server = c("fishbase", "sealifebase"),
  version = "latest",
  ...
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{server}{Either "fishbase" or "sealifebase".}

\item{version}{a version string for the database. See [available_releases()] for details.}
}
\value{
a table of species maturity
}
\description{
maturity
}
\examples{
\dontshow{if (interactive() ) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontrun{
maturity("Oreochromis niloticus")
}
\dontshow{\}) # examplesIf}
}
