% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho_physio.R
\name{morphometrics}
\alias{morphometrics}
\title{morphometrics}
\usage{
morphometrics(species_list = NULL, fields = NULL, query = NULL,
  limit = 200, offset = 0, server = getOption("FISHBASE_API",
  FISHBASE_API), ...)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, the table will be loaded up to the limit. This makes it possible to load an entire table without knowledge of the species in the table, thus avoiding uneccesary API calls.}

\item{fields}{a character vector specifying which fields (columns) should be returned. By default,
all available columns recognized by the parser are returned. This option can be used to limit the amount
of data transfered over the network if only certain columns are needed.}

\item{query}{a named list specifying specific subsets of fields.}

\item{limit}{The maximum number of matches from a single API call (e.g. per species). Function
will warn if this needs to be increased, otherwise can be left as is.}

\item{offset}{The record to start at. Use in combination with \code{limit}
parameter when retrieving more than 5000 results. Default: 0}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use http://fishbase.ropensci.org/sealifebase}

\item{...}{additional arguments to httr::GET}
}
\value{
a table of species morphometrics data
}
\description{
morphometrics
}
\examples{
\dontrun{
morphometrics("Oreochromis niloticus")
}
}

