% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{ping}
\alias{ping}
\title{ping}
\usage{
ping(server = SERVER)
}
\arguments{
\item{server}{base URL to the FishBase API, should leave as default.}
}
\value{
An httr 'response' object.
}
\description{
Check that the FishBase MySQL backend to the API is alos responding
}
\examples{
\dontrun{

## Show server response times
resp <- ping()
resp$times

}
}

