% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refs.R
\name{references}
\alias{references}
\title{references}
\usage{
references(codes = NULL, fields = NULL, server = getOption("FISHBASE_API",
  FISHBASE_API), ...)
}
\arguments{
\item{codes}{One or more Fishbase reference numbers, matching the RefNo 
field}

\item{fields}{a character vector specifying which fields (columns) should be returned. By default,
all available columns recognized by the parser are returned.  Mostly for backwards compatibility as users can subset by column later}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use https://fishbase.ropensci.org/sealifebase}

\item{...}{unused; for backwards compatibility only}
}
\value{
a tibble (data.frame) of reference data
}
\description{
references
}
\examples{
\dontrun{
references(codes = 1)
references(codes = 1:6)
references(codes = 1:6, fields = c('Author', 'Year', 'Title'))
references() # all references
}
}
