% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishbase.R
\docType{package}
\name{rfishbase-package}
\alias{rfishbase}
\alias{rfishbase-package}
\title{The new R interface to Fishbase, v2.0}
\description{
A programmatic interface to http://www.fishbase.org, re-written
based on an accompanying 'RESTful' API. Access tables describing over 30,000
species of fish, their biology, ecology, morphology, and more. This package also
supports experimental access to http://www.sealifebase.org data, which contains
nearly 200,000 species records for all types of aquatic life not covered by
FishBase.'
}
\author{
Carl Boettiger \email{carl@ropensci.org}

Scott Chamberlain \email{scott@ropensci.org}
}
\keyword{package}

