% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/fishbase.R
\docType{package}
\name{rfishbase-package}
\alias{rfishbase}
\alias{rfishbase-package}
\title{The new R interface to Fishbase, v2.0}
\description{
The new R interface to Fishbase, v2.0
}
\author{
Carl Boettiger \email{carl@ropensci.org}

Scott Chamberlain \email{scott@ropensci.org}
}
\keyword{package}

