% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/commonnames.R
\name{sci_to_common}
\alias{sci_to_common}
\title{sci_to_common}
\usage{
sci_to_common(species_list, Language = "English", limit = 1000,
  server = SERVER)
}
\arguments{
\item{Language}{a string specifying the language for the common name, e.g. "English"}

\item{limit}{The maximum number of matches from a single API call (e.g. per species). Function
will warn if this needs to be increased, otherwise can be left as is.}

\item{server}{base URL to the FishBase API, should leave as default.}
}
\value{
The most frequently used common name from the common-names table; see details
}
\description{
Return the best-matching common name given a scientific name (or speccode)
}
\details{
This function will return the common name most frequently used for the species queried (in the
specified language, which defaults to English).  PLEASE NOTE: there is no guarantee that this most frequent
name is indeed the best common name.  Some species have many common names in the same language, reflecting
regional differences (e.g. see \code{commonnames("Salmo Trutta")})
}

