% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxa.R
\docType{data}
\name{sealifebase}
\alias{sealifebase}
\title{A table of all the the species found in SeaLifeBase, including taxonomic
classification and the Species Code (SpecCode) by which the species is
identified in SeaLifeBase}
\description{
A table of all the the species found in SeaLifeBase, including taxonomic
classification and the Species Code (SpecCode) by which the species is
identified in SeaLifeBase
}
\references{
\url{http://www.sealifebase.org}
}
\author{
Carl Boettiger \email{carl@ropensci.org}
}
\keyword{data}
