% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/species_info.R
\docType{data}
\name{species_fields}
\alias{species_fields}
\title{A list of the species_fields available}
\description{
A list of the species_fields available
}
\author{
Carl Boettiger \email{carl@ropensci.org}
}
\references{
\url{FishBase.org}
}
\keyword{data}

