% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_list.R
\name{species_list}
\alias{species_list}
\title{species_list}
\usage{
species_list(Class = NULL, Order = NULL, Family = NULL,
  Subfamily = NULL, Genus = NULL, Species = NULL, SpecCode = NULL,
  SuperClass = NULL, server = getOption("FISHBASE_API", FISHBASE_API))
}
\arguments{
\item{Class}{Request all species in this taxonomic Class}

\item{Order}{Request all species in this taxonomic Order}

\item{Family}{Request all species in this taxonomic Family}

\item{Genus}{Request all species in this taxonomic Genus}

\item{Species}{Request all species in this taxonomic Species}

\item{SpecCode}{Request species name of species matching this SpecCode}

\item{server}{fishbase or}

\item{SubFamily}{Request all species in this taxonomic SubFamily}
}
\description{
Return the a species list given a taxonomic group
}
\examples{
\donttest{
## All species in the Family 
  species_list(Family = 'Scaridae')
## All species in the Genus 
  species_list(Genus = 'Labroides')
}
}
