% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_taxa.R
\name{species_names}
\alias{species_names}
\title{species_names}
\usage{
species_names(server = c("fishbase", "sealifebase"), version = "latest", ...)
}
\arguments{
\item{server}{Either "fishbase" (the default) or "sealifebase"}

\item{version}{the version of the database you want. Will default to the
latest available; see \code{\link[=available_releases]{available_releases()}}.}

\item{...}{for compatibility with previous versions}
}
\value{
the taxa list
}
\description{
Show scientific classification of taxa
Note: species_names() is an alias for load_taxa().
}
