% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_names.R
\name{species_names}
\alias{species_names}
\title{species names}
\usage{
species_names(codes, server = NULL)
}
\arguments{
\item{codes}{a vector of speccodes (e.g. column from a table)}

\item{server}{can be set to either "fishbase" or "sealifebase" to switch between databases. NOTE: it is usually
easier to leave this as NULL and set the source instead using the environmental variable `FISHBASE_API`, e.g.
`Sys.setenv(FISHBASE_API="sealifebase")`.}
}
\value{
A character vector of species names for the SpecCodes
}
\description{
returns species names given FishBase's SpecCodes
}
