% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomy}
\usage{
taxonomy(genus = NULL, species = NULL, ...)
}
\arguments{
\item{genus, species}{(character) genus or species (aka epithet) name. 
pass in either one or both}

\item{...}{additional arguments to \code{\link[httr]{GET}}}

\item{limit}{The maximum number of matches from a single API call (e.g. 
per species). Function will warn if this needs to be increased, otherwise 
can be left as is.}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, 
use https://fishbase.ropensci.org/sealifebase}
}
\description{
Taxonomy
}
\examples{
taxonomy(genus = "Oreochromis", species = "amphimelas")
taxonomy(genus = "Oreochromis")
taxonomy(species = "amphimelas")

taxonomy(genus = "Abrocoma", 
  server = "https://fishbase.ropensci.org/sealifebase")
  
}
