% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{taxonomy}
\alias{taxonomy}
\title{Taxonomy}
\usage{
taxonomy(genus = NULL, species = NULL, ...)
}
\arguments{
\item{genus, species}{(character) genus or species name. 
pass in either one or both
per species). Function will warn if this needs to be increased, otherwise 
can be left as is.}

\item{...}{unused; for backwards compatibility only}
}
\description{
Taxonomy
}
\examples{
taxonomy(genus = "Oreochromis", species = "amphimelas")
taxonomy(genus = "Oreochromis")
taxonomy(species = "amphimelas")

taxonomy(genus = "Abrocoma", 
  server = "https://fishbase.ropensci.org/sealifebase")
  
}
