% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synonyms.R
\name{validate_names}
\alias{validate_names}
\title{validate_names}
\usage{
validate_names(
  species_list,
  server = getOption("FISHBASE_API", "fishbase"),
  version = get_latest_release(),
  db = default_db(),
  ...
)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species"). If empty, a table for all fish will be returned.}

\item{server}{can be set to either "fishbase" or "sealifebase" to switch between databases. NOTE: it is usually
easier to leave this as NULL and set the source instead using the environmental variable `FISHBASE_API`, e.g.
`Sys.setenv(FISHBASE_API="sealifebase")`.}

\item{version}{a version string for the database, will default to the latest release. see [get_releases()] for details.}

\item{db}{the}

\item{...}{unused; for backwards compatibility only}
}
\value{
a string of the validated names
}
\description{
Check for alternate versions of a scientific name and return
the scientific names FishBase recognizes as valid
}
\examples{
\donttest{
validate_names("Abramites ternetzi")
}
}
