% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/synonyms.R
\name{validate_names}
\alias{validate_names}
\title{validate_names}
\usage{
validate_names(species_list, limit = 50, server = SERVER)
}
\arguments{
\item{species_list}{A vector of scientific names (each element as "genus species").}

\item{limit}{The maximum number of matches from a single API call (e.g. per species). Function
will warn if this needs to be increased, otherwise can be left as is.}

\item{server}{base URL to the FishBase API (by default). For SeaLifeBase, use http://fishbase.ropensci.org/sealifebase}
}
\value{
a string of the validated names
}
\description{
Check for alternate versions of a scientific name and return the names FishBase recognizes as valid
}

